address the review issues
* additionally reorder files to avoid duplication
