+++ Gems/SimulationInterfaces/Code/Tests/Common/SimulationInterfaceTestFixture.cpp
@@ -6,45 +6,16 @@
  *
  */
 
-#include "RuntimeTestFixture.h"
-#include "RuntimeTestApplication.h"
+#include "SimulationInterfaceTestFixture.h"
 
+#include <AzCore/Asset/AssetManagerBus.h>
 #include <AzCore/Component/EntityId.h>
-#include <Clients/SimulationEntitiesManager.h>
-#include <Clients/SimulationManager.h>
+#include <AzCore/Component/TransformBus.h>
 #include <SimulationInterfaces/SimulationEntityManagerRequestBus.h>
 #include <gtest/gtest.h>
 
 namespace UnitTest
 {
-    void SimulationInterfaceTestEnvironment::AddGemsAndComponents()
-    {
-        // ROS2 is intentionally absent from dynamic module loading: the gem's Private.Object
-        // already pulls in Gem::ROS2.Static, so libROS2.so is fully resolved in this binary at
-        // link time. dlopen()'ing it again would conflict with the already-resolved symbols.
-        constexpr AZStd::array<AZStd::string_view, 3> requiredGems = { "PhysX5", // required for PhysX Dynamic
-                                                                       "LmbrCentral", // for shapes
-                                                                       "SimulationInterfaces" };
-        AddActiveGems(requiredGems);
-        AddDynamicModulePaths({ "PhysX5.Gem" });
-        AddDynamicModulePaths({ "LmbrCentral" });
-        AddComponentDescriptors(
-            AZStd::initializer_list<AZ::ComponentDescriptor*>{ SimulationInterfaces::SimulationEntitiesManager::CreateDescriptor(),
-                                                               SimulationInterfaces::SimulationManager::CreateDescriptor() });
-        AddRequiredComponents(
-            { SimulationInterfaces::SimulationEntitiesManager::TYPEINFO_Uuid(), SimulationInterfaces::SimulationManager::TYPEINFO_Uuid() });
-    }
-
-    void SimulationInterfaceTestEnvironment::PostSystemEntityActivate()
-    {
-        AZ::UserSettingsComponentRequestBus::Broadcast(&AZ::UserSettingsComponentRequests::DisableSaveOnFinalize);
-    }
-
-    AZ::ComponentApplication* SimulationInterfaceTestEnvironment::CreateApplicationInstance()
-    {
-        return aznew UnitTest::RuntimeTestApplication("SimulationInterfaceTestEnvironment");
-    }
-
     void SimulationInterfaceTestFixture::AddAsset(const AZStd::string& assetPath)
     {
         AZ::Data::AssetInfo info;
@@ -73,7 +44,8 @@ namespace UnitTest
         entity->CreateComponent(AZ::Uuid(PhysXShapeColliderComponentTypeId));
         entity->CreateComponent(AZ::Uuid(SphereShapeComponentTypeId));
         entity->Init();
-        entity->Activate();
+        entity->SetEntityActive(true);
+        entity->ApplyEffectiveActiveState();
         AZ_Assert(entity->GetState() == AZ::Entity::State::Active, "Entity is not active");
 
         // register entity
@@ -106,7 +78,8 @@ namespace UnitTest
                 output, &SimulationInterfaces::SimulationEntityManagerRequests::UnregisterSimulatedBody, entity.second->GetName());
             AZ_Assert(
                 output.IsSuccess(), "Failed to unregister entity from simulation_interfaces, %s", output.GetError().m_errorString.c_str());
-            entity.second->Deactivate();
+            entity.second->SetEntityActive(false);
+            entity.second->ApplyEffectiveActiveState();
         }
         m_entities.clear();
     }
@@ -121,7 +94,8 @@ namespace UnitTest
                 output, &SimulationInterfaces::SimulationEntityManagerRequests::UnregisterSimulatedBody, findIt->second->GetName());
             AZ_Assert(
                 output.IsSuccess(), "Failed to unregister entity from simulation_interfaces, %s", output.GetError().m_errorString.c_str());
-            findIt->second->Deactivate();
+            findIt->second->SetEntityActive(false);
+            findIt->second->ApplyEffectiveActiveState();
             m_entities.erase(findIt);
         }
     }
@@ -161,9 +135,10 @@ namespace UnitTest
 
     void SimulationInterfaceTestFixture::StepPhysics(int numSteps)
     {
+        auto* physicsSystem = AZ::Interface<AzPhysics::SystemInterface>::Get();
+        AZ_Assert(physicsSystem, "Failed to get physics system interface");
         for (int i = 0; i < numSteps; i++)
         {
-            auto* physicsSystem = AZ::Interface<AzPhysics::SystemInterface>::Get();
             physicsSystem->Simulate(1.0f / 60.0f);
         }
     }
@@ -176,6 +151,10 @@ namespace UnitTest
         for (int i = 0; i < numTicks; i++)
         {
             app->Tick();
+            // System tick drains AssetBus (so OnAssetReady notifications fire and assets transition
+            // to Ready) and any SystemTickBus::QueueFunction() lambdas gems posted at Activate time.
+            // Without it, SpawnableEntitiesManager requeues SpawnAllEntitiesCommand forever.
+            app->TickSystem();
         }
     }
